+++ src/main/java/com/course/plazoleta/infraestructure/output/jpa/repository/IRestaurantRepository.java
@@ -3,5 +3,5 @@ package com.course.plazoleta.infraestructure.output.jpa.repository;
 import com.course.plazoleta.infraestructure.output.jpa.entity.RestaurantEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IRestaurantRepository extends JpaRepository<RestaurantEntity, Integer> {
+public interface IRestaurantRepository extends JpaRepository<RestaurantEntity, Long> {
 }
